refactor: move merge_default_tags from CLI to ProviderNormalizer trait
Tag merging is a provider-specific concern (AWS tags), not a CLI concern.
Move the logic into the ProviderNormalizer trait with a default no-op,
and implement it in both AwsccNormalizer and AwsNormalizer. The CLI
now delegates to the provider router instead of doing the merge itself.

Closes #1075
